Fixed Test issues

+++ src/models/mongo/user-mongo-store.js
@@ -33,8 +33,13 @@ export const userMongoStore = {
     }
   },
 
+  // async deleteAll() {
+  //   await User.deleteMany({});
+  // },
+
+  // Delete's all users bar superadmin
   async deleteAll() {
-    await User.deleteMany({});
+    await User.deleteMany({ accountType: { $ne: 'superadmin' } });
   },
 
   async updateUser(userID, updatedUser) {
